Implement Recognition Stage 2 season and episode extraction
Add season/episode syntax parsing, decimal and range support, directory season/cour context, collision protection, deterministic RecognitionEngine output, malformed-path performance hardening, and Stage 2 documentation.

+++ src/Eizo.Metadata.Recognition/RecognitionResult.cs
@@ -7,8 +7,8 @@ public sealed record RecognitionResult(
     MediaKind MediaKind,
     string? Title,
     int? SeasonNumber,
-    int? EpisodeNumber,
-    int? EpisodeEndNumber,
+    decimal? EpisodeNumber,
+    decimal? EpisodeEndNumber,
     int? Year,
     double Confidence,
     IReadOnlyList<RecognitionEvidence> Evidence);
